requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
pipecat-ai>=0.0.40
websockets>=12.0
asyncio>=3.4.3
aiohttp>=3.9.0